Breaking change, changed the api for how "phases" are defined.
They are now defined as "methodActions" and can be either a string
(action type) or a function (action creator). If they are a string, an
action creator function is built which passes params as meta. This will
allow any reducer which handles these actions to access and use any
params that were sent when the action was dispatched.